requirements.txt
autopep8==1.6.0
coloredlogs==15.0.1
confuse==2.0.0
coverage==5.5
deprecation==2.1.0
filetype==1.1.0
humanfriendly==10.0
jellyfish==0.9.0
mediafile==0.9.0
munkres==1.1.4
music-tag==0.4.3
musicbrainzngs==0.7.1
mutagen==1.45.1
packaging==21.3
pycodestyle==2.8.0
pyparsing==3.0.9
pytube==12.1.0
PyYAML==6.0
six==1.16.0
toml==0.10.2
Unidecode==1.3.4
youtube-dl==2021.12.17
